requirements.txt
python-telegram-bot==20.6
transformers
torch
torchvision
docx
nltk
sacrebleu
nest-asyncio